+++ tests/functiontest/conftest.py
@@ -420,13 +420,12 @@ def obj(value, tc_name):
                 if arg_name == "pywbem_object":
                     continue
                 ctor_args[arg_name] = obj(value[arg_name], tc_name)
-            with pytest.warns(None) as rec_warnings:
+            with warnings.catch_warnings(record=True) as rec_warnings:
                 obj_ = ctor_call(**ctor_args)
-            if rec_warnings:
-                for w in rec_warnings.list:
-                    # Ignore DeprecationWarnings, re-issue any others
-                    if w.category != DeprecationWarning:
-                        warnings.warn(w.message, w.category, 1)
+            for w in rec_warnings:
+                # Ignore DeprecationWarnings, re-issue any others
+                if w.category != DeprecationWarning:
+                    warnings.warn(w.message, w.category, 1)
         else:
             obj_ = OrderedDict()
             for key in value: