+++ tests/unittest/utils/pytest_extensions.py
@@ -153,7 +153,7 @@ def simplified_test_function(test_func):
                     assert len(rec_warnings) >= 1, \
                         "Expected warning(s) missing: {}".format(exp_warn_types)
         else:
-            with pytest.warns(None) as rec_warnings:
+            with warnings.catch_warnings(record=True) as rec_warnings:
                 if exp_exc_types:
                     with pytest.raises(exp_exc_types):
                         if condition == 'pdb':
@@ -175,7 +175,7 @@ def simplified_test_function(test_func):
                     # Verify that no warnings have occurred
                     if exp_warn_types is None and rec_warnings:
                         lines = []
-                        for w in rec_warnings.list:
+                        for w in rec_warnings:
                             tup = (w.filename, w.lineno, w.category.__name__,
                                    str(w.message))
                             line = "{t[0]}:{t[1]}: {t[2]}: {t[3]}".format(t=tup)