+++ docs/changes.rst
@@ -77,6 +77,9 @@ Released: not yet
 
 * Test: Upgraded GitHub Actions plugins to use node.js 20.
 
+* Test: Fixed issues resulting from removal of support for pytest.warns(None)
+  in pytest version 8. (issue #3114)
+
 **Enhancements:**
 
 * The 'pywbem.ValueMapping' class now allows controlling what should happen